Move HTTP client/server to new LLVMSupportHTTP
Relocate HTTPClient and HTTPServer from the Debuginfod library to
llvm/Support/HTTP so they can be reused by other components.
