+++ llvm/include/llvm/Support/HTTP/HTTPClient.h
@@ -51,9 +51,7 @@ protected:
 
 /// A reusable client that can perform HTTPRequests through a network socket.
 class HTTPClient {
-#ifdef LLVM_ENABLE_CURL
-  void *Curl = nullptr;
-#endif
+  void *Handle = nullptr;
 
 public:
   HTTPClient();
